Switched from double to float precision

+++ DX11Raytracer/DX11Raytracer/src/DielectricMaterial.h
@@ -7,16 +7,16 @@ namespace gfx
     class DielectricMaterial : public Material
     {
     public:
-        DielectricMaterial(const double indexOfRefraction);
+        DielectricMaterial(const f32 indexOfRefraction);
 
     public:
         virtual const bool Scatter(const Ray& rayIn, const RayHitRecord& rec, Color& attenuation, Color& emission, Ray& scattered, const GBuffer& gBuffer, const uint gBufferIdx, const uint passIteration) const override;
         virtual const bool IsInGBuffer(const uint gBufferIdx) const override;
 
     private:
-        static double SchlickApprox(const double cosine, const double reflectiveIdx);
+        static f32 SchlickApprox(const f32 cosine, const f32 reflectiveIdx);
 
     private:
-        double m_indexOfRefraction;
+        f32 m_indexOfRefraction;
     };
 }
