+++ DX11Raytracer/DX11Raytracer/src/RendererList.h
@@ -19,7 +19,7 @@ namespace gfx
         void Clear();
         void Add(std::shared_ptr<RayReceiver> object) { m_renderers.push_back(object); }
         std::vector<std::shared_ptr<RayReceiver>>& GetRenderers();
-        virtual const bool Hit(const Ray& r, const double t_min, const double t_max, RayHitRecord& rec, const uint gBufferIdx) const override;
+        virtual const bool Hit(const Ray& r, const f32 t_min, const f32 t_max, RayHitRecord& rec, const uint gBufferIdx) const override;
         virtual const bool GetAABB(AABB& aabb) const override;
 
     private:
